jira_api_util/requirements.txt
lorem-text==2.1
python-dotenv==1.0.1